pyproject.toml
[tool.poetry]
name = "edubot-matrix"
version = "0.1.0"
description = "Matrix integration for OpenEdTech Edubot"
authors = ["exciteabletom <tom@digitalnook.net>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.3"
matrix-nio = {version = "^0.20.0", extras = ["e2e"]}
pyyaml = "^6.0"
markdown = "^3.4.1"
#edubot = "^0.3.1"
edubot = {path = "../edubot", develop=true}
feedparser = "^6.0.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
